requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
python-dotenv==1.0.1
exceptiongroup==1.2.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
openai==1.10.0
pip==22.0.2
pydantic==2.6.0
pydantic_core==2.16.1
python-docx==1.1.0
requests==2.31.0
setuptools==59.6.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
